+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 0.10
     - on-screen reader; reads (audio) board messages
+    - encode utf8 characters to '?'
 - 0.9
     - massive refactoring for better modularity and configuration
     - app gets official name: PiBot